+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.1.1
 
 * FluentForms - do_shortcode in Terms & Conditions
+* Disable editor fullscreen mode
+* Archive settings - show all, order by, archive only, don't show "archive" in title
 
 ## 0.1.0